refactor: move ApiProxyCorsComponent in a dedicated folder to be shareable https://gravitee.atlassian.net/browse/APIM-1412
