+++ gravitee-apim-console-webui/src/management/api/cors/api-proxy-cors.component.scss
@@ -1,8 +1,8 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
-@use '../../../../scss/gio-layout' as gio-layout;
+@use 'src/scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-margin-container;
